requirements.txt
streamlit==1.37.1
pandas==2.2.2
pyyaml==6.0.2
qrcode[pil]==7.4.2